ppt_agent_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
redis>=5.0.0
pydantic>=2.0.0
PyJWT>=2.8.0
html2image>=2.0.0
Pillow>=9.0.0
python-pptx>=0.6.21
python-docx>=1.1.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
